test: consolidate crypto tests into common/crypto.test.ts
